Make pci_dma_rw return useful value instead of 0
pci_dma_rw currently returns 0 regardless of whether it failed or not.
Updated to pass through the value from dma_memory_rw. Updated dma_memory_rw to return the MemTxError without casting to bool,
to provide more context. There is a patch(es) addressing this but the changes are not yet upstream.